Update for Python Assignment
For this assignment, I mostly used web scraping to organize and look into the top 100 songs of today. I also defined functions which would allow the user to search up their favorite song or favorite artist, and see if they are a part of the top 100 songs in America today.